Commented all the code and added headless functionality
