+++ UnityProject/Assets/Scripts/Steam.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: b235a8fc0e94a45eaa2f74ec6bcdaac4
+folderAsset: yes
+timeCreated: 1516808961
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
