+++ UnityProject/Assets/Scripts/Steam/SteamClient.cs.meta
